# lib/api-zod/src/generated/types/keyInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ScriptHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KeyInput {
  scriptId: number;
  userId?: number;
  expiresAt?: string;
  whitelisted?: boolean;
}