# lib/api-zod/src/generated/types/panel.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ScriptHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Panel {
  id: number;
  ownerId: number;
  scriptId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  discordServerId?: string | null;
  /** @nullable */
  channelId?: string | null;
  /** @nullable */
  messageId?: string | null;
  requiredRoles?: string[];
  createdAt: string;
}