# lib/api-zod/src/generated/types/panelInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ScriptHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PanelInput {
  scriptId: number;
  /** @minLength 1 */
  name: string;
  description?: string;
  discordServerId?: string;
  channelId?: string;
  requiredRoles?: string[];
}