# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ScriptHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  discordId: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}